+++ src/algorithm/math/bits/exchangeBits.ts
@@ -1,0 +1,7 @@
+// https://leetcode-cn.com/problems/exchange-lcci/
+// 金典-05.07
+export function exchangeBits(num: number): number {
+  const oddMask = 0b0101010101010101010101010101010101010101;
+  const evenMask = 0b1010101010101010101010101010101010101010;
+  return ((num & oddMask) << 1) | ((num & evenMask) >> 1);
+}